Init gRPC service and RetrievePcap call
This commit adds a gRPC service to Stenographer and implements a call to retrieve PCAP via stenoread. Given the sensitivity of packet data, the gRPC channel is designed to only support encryption with client authentication. Due to the scope of many environments (i.e. there are multiple servers on which Stenographer is running), the gRPC service does not share SSL certificates with the local Stenographer utility (perhaps these can be merged in the future). Note that this commit explicitly does not add any client-side gRPC functionality. Below is a summary of the additions and changes:
- Added rpc/rpc.go
- Added protobuf/steno.proto
- Changed config/config.go to support Rpc configuration
- Changed stenographer.go to check for Rpc configuration and (if found) run gRPC server as a goroutine
- Updated README and DESIGN docs to describe Rpc functionality