Add check for empty PCAP files and removeFile func
This commit adds a logic check to the RetrievePcap call that checks if the stenoread-parsed PCAP is exactly 24 bytes (this means the PCAP contains no packets); if it doesn’t, then nothing is streamed to the client. This also identified an edge case where we may want to delete PCAP files from disk without having streamed PCAP data back to the client — to resolve that, the func removeFile was added and is now used throughout the RetrievePcap call.